refactor: rename Order component to Cart
